Revised the way the pipeline works.
Rather than using an input text file to describe the pipeline flow,
what was in rtsprocess.py is now rewritten to work as a python package
(mwa_pipe.py) which enables a python script to replace the text file.
This allows greater flexibility as it is possible to do other tasks in
between the RTS-related processing.