Alternate data path and noise-weighted image integration
Added the ability to set an alternate data path for data retrieved from
the archive (useful it a common data pool is used by many users but a
local path is needed for RTS-generated output files). Also added an
alternate version of the image integration script that now checks for
Stokes V images and uses the noise in those images as a weighting
factor when integrating over multiple images.